Add generic ExperimentConfig for new experiments
